refactor(frontend): split vram upload queue
